refactor: sever stay-in-core rig edges (homeboy-rig extraction prep)
Prep for extracting the rig subsystem (~15.3k LOC) into its own crate. Core
files that stay in core referenced rig types + one toolchain behavior.

1. RigStateSnapshot (5 refs in extension/trace + bench) was ALREADY re-exported
   from homeboy-lifecycle-contract; repointed those uses directly to the
   contract — zero new infra.
2. Trace*Spec types (TraceDependencySpec, TracePublicPreviewSpec,
   TracePreviewAssetFanoutSpec, TracePublicPreviewMode,
   TraceNativePublicPreviewSpec) — moved the 5 self-contained data types to a
   new homeboy-rig-contract crate (below core). extension/trace reads them from
   the contract; rig re-exports them from trace.rs so its internals are
   unchanged.
3. extension/execution.rs called rig::toolchain::command_step_path() to build
   the exec-env PATH. Inverted behind a RigToolchainProvider hook.

The only remaining core->rig ref is trace_experiment.rs (rig-experiment logic
misfiled at core top-level), which moves WITH rig at the git-mv. rig has no
other feature-crate deps (only homeboy-lab-runner-contract, below core).

cargo check --workspace --tests: green.
