+++ crates/homeboy-core/Cargo.toml
@@ -32,6 +32,7 @@ homeboy-refactor-contract = { version = "0.1.0", path = "../homeboy-refactor-con
 homeboy-code-audit = { version = "0.1.0", path = "../homeboy-code-audit" }
 homeboy-extension-contract = { version = "0.1.0", path = "../homeboy-extension-contract" }
 homeboy-lifecycle-contract = { version = "0.1.0", path = "../homeboy-lifecycle-contract" }
+homeboy-rig-contract = { version = "0.1.0", path = "../homeboy-rig-contract" }
 homeboy-engine-primitives = { version = "0.1.0", path = "../homeboy-engine-primitives" }
 homeboy-lab-contract = { version = "0.1.0", path = "../homeboy-lab-contract" }
 homeboy-lab-runner-contract = { version = "0.1.0", path = "../homeboy-lab-runner-contract" }
